refactor: Main, Modal Component 파일 분리 및 index.js 를 이용하여 Component export 통일
