+++ src/App.js
@@ -1,8 +1,8 @@
 import './index.css';
 import './css/index.css';
 import React, { Component } from 'react';
-import Main from './components/Main';
-import Modal from './components/Modal';
+import { Main } from './components/Main';
+import { Modal } from './components/Modal';
 import { LOTTO_VALUE, LOTTO_PRICE } from './constants';
 import { getAnnouncementDate } from './utils/lottoUtils';
